feat:implement isPrimary mutation for emails
