+++ packages/apps/frontera/src/routes/finder/src/components/ContactPreviewCard/ContactPreviewCard.tsx
@@ -1,5 +1,5 @@
+import { useState } from 'react';
 import { useSearchParams } from 'react-router-dom';
-import { useRef, Fragment, useState } from 'react';
 
 import set from 'lodash/set';
 import { useKeyBindings } from 'rooks';
@@ -10,44 +10,26 @@ import { cn } from '@ui/utils/cn';
 import { Input } from '@ui/form/Input';
 import { flags } from '@ui/media/flags';
 import { Avatar } from '@ui/media/Avatar';
-import { Plus } from '@ui/media/icons/Plus';
 import { DateTimeUtils } from '@utils/date';
 import { Tag01 } from '@ui/media/icons/Tag01';
 import { Spinner } from '@ui/feedback/Spinner';
-import { Mail02 } from '@ui/media/icons/Mail02';
 import { Star06 } from '@ui/media/icons/Star06';
-import { Button } from '@ui/form/Button/Button';
-import { getTimezone } from '@utils/getTimezone';
 import { IconButton } from '@ui/form/IconButton';
+import { getTimezone } from '@utils/getTimezone';
 import { useStore } from '@shared/hooks/useStore';
-import { Archive } from '@ui/media/icons/Archive';
 import { Tags } from '@organization/components/Tabs';
 import { Tooltip } from '@ui/overlay/Tooltip/Tooltip';
-import { TextInput } from '@ui/media/icons/TextInput';
-import { PlusCircle } from '@ui/media/icons/PlusCircle';
 import { getFormattedLink } from '@utils/getExternalLink';
 import { Tag, Social, TableViewType } from '@graphql/types';
-import { DotsVertical } from '@ui/media/icons/DotsVertical';
 import { LinkedInSolid02 } from '@ui/media/icons/LinkedInSolid02';
-import { Menu, MenuItem, MenuList, MenuButton } from '@ui/overlay/Menu/Menu';
-import { EmailValidationMessage } from '@organization/components/Tabs/panels/PeoplePanel/ContactCard/EmailValidationMessage';
-import {
-  Modal,
-  ModalBody,
-  ModalClose,
-  ModalPortal,
-  ModalFooter,
-  ModalOverlay,
-  ModalCloseButton,
-  ModalFeaturedHeader,
-  ModalFeaturedContent,
-} from '@ui/overlay/Modal';
+
+import { EmailsSection } from './components';
+import { EnrichContactModal } from './components/EnrichContactModal';
 
 export const ContactPreviewCard = observer(() => {
   const store = useStore();
   const [searchParams] = useSearchParams();
   const [isEditName, setIsEditName] = useState(false);
-  const [isLoading, setIsLoading] = useState(false);
   const contactId = store.ui.focusRow;
   const preset = searchParams?.get('preset');
   const tableViewDef = store.tableViewDefs.getById(preset ?? '1');
@@ -80,8 +62,6 @@ export const ContactPreviewCard = observer(() => {
   const timezone = city
     ? cityTimezone.lookupViaCity(city)?.[0]?.timezone
     : null;
-
-  const validationDetails = contact?.value.emails?.[0]?.emailValidationDetails;
 
   const fromatedUrl = contact?.value?.socials?.[0]?.url.replace(
     'https://www.',
@@ -158,7 +138,7 @@ export const ContactPreviewCard = observer(() => {
       {store.ui.contactPreviewCardOpen && (
         <div
           data-state={store.ui.contactPreviewCardOpen ? 'open' : 'closed'}
-          className='data-[state=open]:animate-slideLeftAndFade data-[state=closed]:animate-slideRightAndFade flex flex-col absolute right-[12px] -top-[-53px] p-4 max-w-[390px] min-w-[350px] border border-gray-200 rounded-lg z-50 bg-white'
+          className='data-[state=open]:animate-slideLeftAndFade data-[state=closed]:animate-slideRightAndFade flex flex-col absolute right-[12px] -top-[-53px] p-4 max-w-[390px] min-w-[350px] border border-gray-200 rounded-lg z-[1] bg-white'
         >
           <div className='flex justify-between items-start'>
             <Avatar
@@ -169,7 +149,7 @@ export const ContactPreviewCard = observer(() => {
               src={src || undefined}
             />
             {!userBeenEnriched && (
-              <Tooltip label='Enrich this contact'>
+              <Tooltip asChild label='Enrich this contact'>
                 {!requestedEnrichment ? (
                   <IconButton
                     size='xxs'
@@ -257,162 +237,7 @@ export const ContactPreviewCard = observer(() => {
             )}
           </div>
           <div className='flex justify-between gap-1 w-full mb-4 flex-col'>
-            <div className='flex items-center justify-between w-full text-sm group/menu'>
-              <div className='flex items-center gap-2'>
-                <Mail02 className='mt-[1px] text-gray-500' />
-                <span className='text-gray-500'>Emails</span>
-                {contact?.value.emails.length === 0 && (
-                  <span className='text-gray-400 ml-[57px]'>No emails yet</span>
-                )}
-              </div>
-              {company && (
-                <div className='flex items-center gap-2'>
-                  <Menu>
-                    <MenuButton>
-                      <Tooltip
-                        align='end'
-                        side='bottom'
-                        label={'Add new email'}
-                      >
-                        <IconButton
-                          size='xxs'
-                          variant='ghost'
-                          icon={<Plus />}
-                          aria-label='add new email'
-                        />
-                      </Tooltip>
-                    </MenuButton>
-                    <MenuList>
-                      <MenuItem
-                        className='group/find-email '
-                        onClick={() => {
-                          setIsLoading(true);
-                          contact
-                            ?.findEmail()
-                            .finally(() =>
-                              setTimeout(() => setIsLoading(false), 60000),
-                            );
-                        }}
-                      >
-                        <div className='flex items-center gap-1'>
-                          <Star06 className='group-hover/find-email:text-gray-700 text-gray-500' />
-                          <span className='max-w-[150px] text-ellipsis overflow-hidden whitespace-nowrap'>{`Find email at ${company}`}</span>
-                        </div>
-                      </MenuItem>
-                      <MenuItem
-                        className='group/add-email'
-                        onClick={() => {
-                          store.ui.setSelectionId(contact.value.emails.length);
-                          contact.update(
-                            (c) => {
-                              c.emails.push({
-                                id: crypto.randomUUID(),
-                                email: '',
-                                appSource: '',
-                                contacts: [],
-                                createdAt: new Date().toISOString(),
-                                updatedAt: new Date().toISOString(),
-                                // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                              } as any);
-
-                              return c;
-                            },
-                            { mutate: false },
-                          );
-                          store.ui.commandMenu.setContext({
-                            ids: [contact.value.id],
-                            entity: 'Contact',
-                            property: 'email',
-                          });
-                          store.ui.commandMenu.setType('EditEmail');
-                          store.ui.commandMenu.setOpen(true);
-                        }}
-                      >
-                        <PlusCircle className='text-gray-500 group-hover/add-email:text-gray-700' />
-                        Add new email
-                      </MenuItem>
-                    </MenuList>
-                  </Menu>
-                  {isLoading && (
-                    <Tooltip label={`Finding email at ${company}...`}>
-                      <div>
-                        <Spinner
-                          size='sm'
-                          label='finding email'
-                          className='text-gray-400 fill-gray-700'
-                        />
-                      </div>
-                    </Tooltip>
-                  )}
-                </div>
-              )}
-            </div>
-            <div className='ml-6'>
-              {contact?.value.emails.map((email, idx) => (
-                <Fragment key={email.id}>
-                  <div className=' flex items-center justify-between '>
-                    <div key={email.id} className='flex items-center '>
-                      <span className='text-sm'>{email.email}</span>
-                    </div>
-                    <div className='flex items-center gap-2'>
-                      {email && (
-                        <EmailValidationMessage
-                          email={email?.email || ''}
-                          validationDetails={validationDetails}
-                        />
-                      )}
-                      <Menu>
-                        <MenuButton>
-                          <IconButton
-                            size='xxs'
-                            variant='ghost'
-                            icon={<DotsVertical />}
-                            aria-label='add new email'
-                          />
-                        </MenuButton>
-                        <MenuList>
-                          <MenuItem
-                            className='group/edit-email'
-                            onClick={() => {
-                              store.ui.setSelectionId(idx);
-                              store.ui.commandMenu.setType('EditEmail');
-                              store.ui.commandMenu.setContext({
-                                ids: [contact.value.id],
-                                entity: 'Contact',
-                                property: 'email',
-                              });
-                              store.ui.commandMenu.setOpen(true);
-                            }}
-                          >
-                            <div className='flex items-center gap-2'>
-                              <TextInput className='group-hover/edit-email:text-gray-700 text-gray-500' />
-                              <span>Edit email</span>
-                            </div>
-                          </MenuItem>
-                          <MenuItem
-                            className='group/archive-email'
-                            onClick={() => {
-                              contact.update(
-                                (c) => {
-                                  c.emails.splice(idx, 1);
-
-                                  return c;
-                                },
-                                { mutate: false },
-                              );
-                              contact.updateEmail(email?.email ?? '');
-                            }}
-                          >
-                            <Archive className='text-gray-500 group-hover/archive-email:text-gray-700' />
-                            Archive email
-                          </MenuItem>
-                        </MenuList>
-                      </Menu>
-                    </div>
-                  </div>
-                </Fragment>
-              ))}
-            </div>
+            <EmailsSection contactId={contactId} />
           </div>
 
           <div className='flex justify-between gap-1 w-full mb-4'>
@@ -509,128 +334,3 @@ export const ContactPreviewCard = observer(() => {
     </>
   );
 });
-
-const EnrichContactModal = observer(
-  ({
-    isModalOpen = false,
-    onClose,
-    contactId,
-  }: {
-    onClose: () => void;
-    isModalOpen: boolean;
-    contactId: string | number;
-  }) => {
-    const store = useStore();
-    const hasSubmitedRef = useRef(false);
-    const [linkedin, setLinkedin] = useState(
-      () => store.contacts.value.get(String(contactId))?.value.socials[0]?.url,
-    );
-    const [validation, setValidation] = useState<Record<'linkedin', boolean>>({
-      linkedin: false,
-    });
-
-    const contactStore = store.contacts.value.get(String(contactId));
-
-    const validate = () => {
-      setValidation(() => ({
-        linkedin: !linkedin,
-      }));
-
-      return linkedin;
-    };
-
-    const reset = () => {
-      setLinkedin('');
-      setValidation({
-        linkedin: false,
-      });
-      hasSubmitedRef.current = false;
-    };
-
-    const handleSubmit = () => {
-      hasSubmitedRef.current = true;
-
-      if (!validate()) return;
-
-      contactStore?.addSocial(linkedin || '', {
-        onSuccess: () => {
-          onClose();
-          reset();
-        },
-      });
-    };
-
-    return (
-      <Modal
-        open={isModalOpen}
-        onOpenChange={(open) => {
-          if (!open) {
-            reset();
-            onClose();
-          }
-        }}
-      >
-        <ModalPortal>
-          <ModalOverlay className='z-[999]' />
-          <ModalFeaturedContent className='z-[9999]'>
-            <ModalFeaturedHeader>
-              <p className='text-lg font-semibold mb-1'>
-                What’s this contact’s LinkedIn?
-              </p>
-              <p className='text-sm'>
-                To enrich this contact, we need their LinkedIn URL
-              </p>
-            </ModalFeaturedHeader>
-            <ModalCloseButton />
-            <ModalBody className='flex flex-col gap-4'>
-              <div className='flex flex-col'>
-                <Input
-                  id='linkedin'
-                  value={linkedin}
-                  placeholder='LinkedIn profile link'
-                  className={cn(validation.linkedin && 'border-error-500')}
-                  onChange={(e) => {
-                    setLinkedin(e.target.value);
-                  }}
-                  onKeyDown={(e) => {
-                    if (e.key === 'Escape') {
-                      onClose();
-                    }
-                    e.stopPropagation();
-                  }}
-                />
-                {validation.linkedin && (
-                  <p className='text-sm text-error-500 mt-1'>
-                    One does not simply skip LinkedIn
-                  </p>
-                )}
-              </div>
-            </ModalBody>
-            <ModalFooter className='flex gap-3'>
-              <ModalClose className='w-full'>
-                <Button className='w-full'>Cancel</Button>
-              </ModalClose>
-
-              <Button
-                className='w-full'
-                colorScheme='primary'
-                onClick={handleSubmit}
-                loadingText='Creating contact'
-                isLoading={store.contacts.isLoading}
-                rightSpinner={
-                  <Spinner
-                    size='sm'
-                    label='loading'
-                    className='text-primary-500 fill-primary-200'
-                  />
-                }
-              >
-                Enrich contact
-              </Button>
-            </ModalFooter>
-          </ModalFeaturedContent>
-        </ModalPortal>
-      </Modal>
-    );
-  },
-);
