+++ packages/apps/frontera/src/routes/finder/src/components/ContactPreviewCard/components/EmailsSection/index.ts
@@ -1,0 +1,1 @@
+export * from './EmailsSection';
