+++ packages/apps/frontera/src/routes/finder/src/components/ContactPreviewCard/components/EnrichContactModal/index.ts
@@ -1,0 +1,1 @@
+export * from './EnrichContactModal';
